스케줄 칼로리 수정

+++ fitory/src/main/java/com/Fitory/fitory/dto/detailfood.java
@@ -9,7 +9,7 @@ import java.time.LocalTime;
 @Setter
 public class detailfood {
 
-    private int calories;
+    private double calories;
 
 
     private double protein;
